add setup to server_test

+++ src/solidity/lib/SignatureVerifier.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.8.13;
 
 library SignatureVerifier {
     struct TimedSignature {
-        uint256 validFor;
+        uint64 validFor;
         bytes32 messageHash;
         bytes signature;
         address signer;
